+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-BzkfMk5f.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-DmPk8dFN.js";import{_ as a}from"./mermaid.core-Doyb5Ydy.js";import"./chunk-5VM5RSS4-IUfKd8Mh.js";import"./chunk-XXDRQBXY-WmB-U0pM.js";import"./chunk-POPQ4Y6H-Dq21YUXF.js";import"./chunk-F27PBJKO-H3JOi8-f.js";import"./channel-T1ZsYZjD.js";import"./index-vgWwsZzc.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};