+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DtWRqXDV.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-D9OAkJQT.js";import{_ as a}from"./mermaid.core-BFYyWtHM.js";import"./chunk-5VM5RSS4-aR8E6glA.js";import"./chunk-XXDRQBXY-Q8a-_LCl.js";import"./chunk-POPQ4Y6H-D5vux6cP.js";import"./chunk-F27PBJKO-JJ--2qby.js";import"./channel-DznPIEF6.js";import"./index-C3Whtis9.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};